backend-py/pyproject.toml
[project]
name = "backend-py"
version = "0.1.0"
description = "Agent-based RAG backend for EleccionesCR 2026"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "qdrant-client>=1.12.1",
    "sentence-transformers>=3.3.1",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "langchain>=0.3.13",
    "langgraph>=0.2.60",
    "langchain-qdrant>=0.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "httpx>=0.27.0",  # for testing FastAPI
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
